feat(codemods): call build:cli on prepack. Move cli implementation to src/
