+++ packages/codemods/package.json
@@ -9,11 +9,10 @@
     "directory": "packages/codemods"
   },
   "type": "module",
-  "bin": {
-    "codemods": "codemods"
-  },
+  "bin": "dist/index.js",
   "scripts": {
-    "build:cli": "bun build bin/index.ts --compile --target node --outfile codemods --sourcemap",
+    "build:cli": "bun build src/cli/index.ts --target node --outdir dist --sourcemap",
+    "prepack": "pnpm run build:cli",
     "lint": "eslint . --quiet --cache --cache-strategy=content"
   },
   "exports": {
