+++ packages/codemods/src/cli/index.ts
@@ -1,6 +1,6 @@
 import { program } from 'commander';
 
-import { version } from '../package.json' with { type: 'json' };
+import { version } from '../../package.json' with { type: 'json' };
 import { createApplyCommand } from './apply.js';
 import { createListCommand } from './list.js';
